Extract raw public key tests
